removed conflicts in signup page
