--- frontend/src/components/Login/Login.css
@@ -1,4 +1,0 @@
-.showAlert {
-    color: red;
-  }
-
