--- frontend/src/components/SignUp/Signup.css
@@ -1,4 +1,0 @@
-.showAlert {
-    color: red;
-  }
-
